+++ requirements.txt
@@ -1,3 +1,5 @@
 python-dotenv~=1.0.1
 ujson~=5.10.0
 httpx~=0.27.0
+beautifulsoup4~=4.12.3
+lxml~=5.3.0
